Refactor User Model and Add Age Calculation Method
This commit refactors the  model in the  file by updating its attribute names and types. The following changes were made:

- Changed  attribute to
- Added  attribute of type
- Changed  attribute to  of type
- Added  attribute of type
- Added  attribute of type  with default value of

Additionally, a new method  was added to the  model. This method calculates the age of the user based on the  attribute.

These changes improve the clarity and organization of the  model and provide a convenient method for calculating user age.

Fixes #<issue_number>